pyproject.toml
[project]
name = "cat-behavior-chatbot"
version = "0.1.0"
description = "Domain-restricted Q&A chatbot for cat behavior using TinyLlama + FastAPI"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
py-modules = ["app", "eval"]

[tool.uv]
package = true
